Change `SharedTestUtilities` to `target` from `testTarget`
